Page, Card, Step creation
moving steps and handlers to Page component
adding useCallback
