+++ src/App.jsx
@@ -1,11 +1,11 @@
-import StepTracker from './components/StepTracker'
+import Page from './components/Page';
 import './App.css'
 
 function App() {
 
   return (
     <>
-      <StepTracker />
+      <Page />
     </>
   )
 }
