+++ src/components/Card.jsx
@@ -1,0 +1,11 @@
+const Card = ({activeStep, handleNext, handleBack}) => {
+    return (
+        <>
+            <div>card content will go here</div>
+            <button onClick={handleNext}>next</button>
+            <button onClick={handleBack}>back</button>
+        </>
+    )
+}
+
+export default Card;
